+++ src/components/molecules/Input/index.test.tsx
@@ -2,6 +2,11 @@ import React from 'react';
 import {render, fireEvent} from '@testing-library/react-native';
 import Input, {InputProps} from './index';
 
+jest.mock('react-native/Libraries/Utilities/Platform', () => ({
+	OS: 'web',
+	select: () => null,
+}));
+
 describe('Input Component', () => {
 	describe('it does not render', () => {
 		it('as it is amountTotal type and no valid totalValue is passed', () => {
@@ -16,57 +21,57 @@ describe('Input Component', () => {
 
 	describe('keyboardTypes', () => {
 		it('should apply the correct keyboardType for currency', () => {
-			const {getByPlaceholderText} = render(<Input placeholder="Enter amount" type="currency" />);
-			const input = getByPlaceholderText('Enter amount');
+			const {getByTestId} = render(<Input placeholder="Enter amount" type="currency" />);
+			const input = getByTestId('input');
 			expect(input.props.keyboardType).toBe('numeric');
 		});
 
 		it('should apply the correct keyboardType for number', () => {
-			const {getByPlaceholderText} = render(<Input placeholder="Enter number" type="number" />);
-			const input = getByPlaceholderText('Enter number');
+			const {getByTestId} = render(<Input placeholder="Enter number" type="number" />);
+			const input = getByTestId('input');
 			expect(input.props.keyboardType).toBe('numeric');
 		});
 
 		it('should apply the correct keyboardType for text', () => {
-			const {getByPlaceholderText} = render(<Input placeholder="Enter text" type="text" />);
-			const input = getByPlaceholderText('Enter text');
+			const {getByTestId} = render(<Input placeholder="Enter text" type="text" />);
+			const input = getByTestId('input');
 			expect(input.props.keyboardType).toBe('default');
 		});
 
 		it('should apply the correct keyboardType for email', () => {
-			const {getByPlaceholderText} = render(<Input placeholder="Enter email" type="email" />);
-			const input = getByPlaceholderText('Enter email');
+			const {getByTestId} = render(<Input placeholder="Enter email" type="email" />);
+			const input = getByTestId('input');
 			expect(input.props.keyboardType).toBe('email-address');
 		});
 
 		it('should apply the correct keyboardType for phone', () => {
-			const {getByPlaceholderText} = render(<Input placeholder="Enter phone" type="phone" />);
-			const input = getByPlaceholderText('Enter phone');
+			const {getByTestId} = render(<Input placeholder="Enter phone" type="phone" />);
+			const input = getByTestId('input');
 			expect(input.props.keyboardType).toBe('phone-pad');
 		});
 	});
 
 	describe('should render correctly', () => {
 		it('as default props are passed', () => {
-			const {getByPlaceholderText} = render(<Input placeholder="Enter text" type="text" />);
-			const input = getByPlaceholderText('Enter text');
+			const {getByTestId} = render(<Input placeholder="Enter text" type="text" />);
+			const input = getByTestId('input');
 			expect(input).toBeTruthy();
 		});
 
 		it('and call changeTextCb when text is changed', () => {
 			const onChangeTextMock = jest.fn();
-			const {getByPlaceholderText} = render(
+			const {getByTestId} = render(
 				<Input placeholder="Enter text" type="text" onChangeText={onChangeTextMock} />
 			);
-			const input = getByPlaceholderText('Enter text');
+			const input = getByTestId('input');
 			fireEvent.changeText(input, 'new text');
 			expect(onChangeTextMock).toHaveBeenCalledWith('new text');
 		});
 
 		it('and do not call changeTextCb when text is changed as there is not one', () => {
 			const onChangeTextMock = jest.fn();
-			const {getByPlaceholderText} = render(<Input placeholder="Enter text" type="text" />);
-			const input = getByPlaceholderText('Enter text');
+			const {getByTestId} = render(<Input placeholder="Enter text" type="text" />);
+			const input = getByTestId('input');
 			fireEvent.changeText(input, 'new text');
 			expect(onChangeTextMock).toHaveBeenCalledTimes(0);
 		});
@@ -92,13 +97,21 @@ describe('Input Component', () => {
 		});
 
 		it('when keyboardType is default as no valid combination of variant nor type is passed', () => {
-			const {getByPlaceholderText} = render(
+			const {getByTestId} = render(
 				<Input placeholder="Enter text" {...({variant: 'test'} as unknown as InputProps)} />
 			);
 
-			const input = getByPlaceholderText('Enter text');
+			const input = getByTestId('input');
 
 			expect(input.props.keyboardType).toBe('default');
+		});
+
+		it('with no placeholder', () => {
+			const {getByTestId} = render(<Input type="number" value={'Hola'} />);
+
+			const input = getByTestId('input');
+
+			expect(input.props.keyboardType).toBe('numeric');
 		});
 	});
 });