+++ src/theme/typography.ts
@@ -1,14 +1,16 @@
 import {moderateScale, scaledForDevice} from 'scale';
 
+const scaleSize = (size: number) => scaledForDevice(size, moderateScale);
+
 export type TypographyItem = {
-	weight: '100' | '200' | '300' | '400' | '500' | '600' | '700' | '800' | '900';
-	size: number;
+	fontWeight: '100' | '200' | '300' | '400' | '500' | '600' | '700' | '800' | '900';
+	fontSize: number;
 	lineHeight: number;
-	spacing?: number;
+	letterSpacing?: number;
 };
 
 export type Typography = {
-	display: TypographyItem;
+	display: {medium: TypographyItem};
 	heading: {
 		large: TypographyItem;
 		medium: TypographyItem;
@@ -36,30 +38,94 @@ export type Typography = {
 };
 
 const typography: Typography = {
-	display: {weight: '400', size: scaledForDevice(42, moderateScale), lineHeight: 50},
+	display: {
+		medium: {
+			fontSize: scaleSize(42),
+			fontWeight: '400',
+			lineHeight: 50,
+		},
+	},
 	heading: {
-		large: {weight: '500', size: scaledForDevice(34, moderateScale), lineHeight: 40},
-		medium: {weight: '500', size: scaledForDevice(26, moderateScale), lineHeight: 32},
-		small: {weight: '400', size: scaledForDevice(24, moderateScale), lineHeight: 28},
+		large: {
+			fontSize: scaleSize(34),
+			fontWeight: '500',
+			lineHeight: 40,
+		},
+		medium: {
+			fontSize: scaleSize(26),
+			fontWeight: '500',
+			lineHeight: 32,
+		},
+		small: {
+			fontSize: scaleSize(24),
+			fontWeight: '400',
+			lineHeight: 28,
+		},
 	},
 	title: {
-		large: {weight: '400', size: scaledForDevice(20, moderateScale), lineHeight: 24},
-		medium: {weight: '700', size: scaledForDevice(18, moderateScale), lineHeight: 22},
-		small: {weight: '700', size: scaledForDevice(14, moderateScale), lineHeight: 16},
+		large: {
+			fontSize: scaleSize(20),
+			fontWeight: '400',
+			lineHeight: 24,
+		},
+		medium: {
+			fontSize: scaleSize(18),
+			fontWeight: '700',
+			lineHeight: 22,
+		},
+		small: {
+			fontSize: scaleSize(14),
+			fontWeight: '700',
+			lineHeight: 16,
+		},
 	},
 	label: {
-		large: {weight: '500', size: scaledForDevice(16, moderateScale), lineHeight: 18},
-		medium: {weight: '500', size: scaledForDevice(14, moderateScale), lineHeight: 16},
-		small: {weight: '500', size: scaledForDevice(12, moderateScale), lineHeight: 14},
+		large: {
+			fontSize: scaleSize(16),
+			fontWeight: '500',
+			lineHeight: 18,
+		},
+		medium: {
+			fontSize: scaleSize(14),
+			fontWeight: '500',
+			lineHeight: 16,
+		},
+		small: {
+			fontSize: scaleSize(12),
+			fontWeight: '500',
+			lineHeight: 14,
+		},
 	},
 	body: {
-		large: {weight: '400', size: scaledForDevice(16, moderateScale), lineHeight: 20},
-		medium: {weight: '400', size: scaledForDevice(14, moderateScale), lineHeight: 18},
-		small: {weight: '400', size: scaledForDevice(12, moderateScale), lineHeight: 16},
+		large: {
+			fontSize: scaleSize(16),
+			fontWeight: '400',
+			lineHeight: 20,
+		},
+		medium: {
+			fontSize: scaleSize(14),
+			fontWeight: '400',
+			lineHeight: 18,
+		},
+		small: {
+			fontSize: scaleSize(12),
+			fontWeight: '400',
+			lineHeight: 16,
+		},
 	},
 	overline: {
-		large: {weight: '500', size: scaledForDevice(14, moderateScale), lineHeight: 16, spacing: 1},
-		small: {weight: '500', size: scaledForDevice(12, moderateScale), lineHeight: 14, spacing: 0.7},
+		large: {
+			fontSize: scaleSize(14),
+			fontWeight: '500',
+			lineHeight: 16,
+			letterSpacing: 1,
+		},
+		small: {
+			fontSize: scaleSize(12),
+			fontWeight: '500',
+			lineHeight: 14,
+			letterSpacing: 0.7,
+		},
 	},
 };