+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Disable Button colors.
+- Fixed Input styles - [JUIP-166](https://janiscommerce.atlassian.net/browse/JUIP-166)
 
 ## [1.14.0] - 2025-02-05